adding what i have so far
Sample.html is how i want the information to be displayed at the current time, will need to change it later
searchprocess.py is the ELSA query, needs to be finished still. need to make it more dynamic. also need to finish making it output the information that it gets.